+++ ggshield/verticals/hmsl/crypto.py
