+++ ggshield/verticals/hmsl/utils.py
@@ -15,7 +15,7 @@ from ggshield.core.errors import (
     MissingTokenError,
     UnknownInstanceError,
 )
-from ggshield.hmsl.client import HMSLClient
+from ggshield.verticals.hmsl.client import HMSLClient
 
 
 logger = logging.getLogger(__name__)
