+++ tests/unit/cmd/hmsl/test_fingerprint.py
@@ -5,8 +5,8 @@ from click.testing import CliRunner
 
 from ggshield.cmd.hmsl.decrypt import load_mapping
 from ggshield.cmd.main import cli
-from ggshield.hmsl.client import PREFIX_LENGTH
-from ggshield.hmsl.crypto import hash_string
+from ggshield.verticals.hmsl.client import PREFIX_LENGTH
+from ggshield.verticals.hmsl.crypto import hash_string
 from tests.unit.conftest import assert_invoke_exited_with, assert_invoke_ok
